requirements.txt
django==1.7.1
psycopg2
django-genericadmin==0.6.1
django-extensions == 1.3.8
